Convert to Python 3 on setup
Include psycopg2 custom b-fixer to convert strings annotated with b()
to bytes. Note that because psycopg2-ctypes is for python >= 2.6 only
the b() function could be dropped and b"" could be used directly. We can
do this refactor later if needed (albeit it would make keeping the code
in sync with psycopg2 a little bit painful... we'll think about that).